some debugging dif to version that crashes

+++ src/rat.cc
@@ -28,7 +28,7 @@ void Rat::packets_received( const vector< Packet > & packets ) {
   const Whisker & current_whisker( _whiskers.use_whisker( _memory, _track ) );
   double srtt = _memory.srtt();
   _the_window = current_whisker.window( _the_window );
-  _intersend_time = ( current_whisker.intersend()) * (srtt/_the_window);
+  _intersend_time = ( current_whisker.intersend()) * (srtt/double(_the_window));
 }
 
 void Rat::reset( const double & )
